feat: add `TypeAssertionError` issue tree
`TypeAssertionError`s now have an `issueTree` property and an
`issueTreeNode` method. Their `toJSON` method now returns a JSON
representation of the `issueTree`.

The `_TypeAssertionIssueTree` `Asserter` has also been added.

The `TypeAssertionError` constructor now wraps inner error value names
with brackets and quotes in the `message`, or prefixes them with a dot,
as appropriate.